refactor(P3): 불필요한 클래스 삭제
