+++ src/main/java/onboarding/Problem3.java
@@ -1,11 +1,10 @@
 package onboarding;
 
-import onboarding.problem3.Check369;
 import onboarding.problem3.ThreeSixNineGame;
 
 public class Problem3 {
     public static int solution(int number) {
-        ThreeSixNineGame threeSixNineGame = new ThreeSixNineGame(new Check369());
+        ThreeSixNineGame threeSixNineGame = new ThreeSixNineGame();
         return threeSixNineGame.do369(number);
     }
 }
